refactor: move backup all plugins
